pyproject.toml
[project]
name = "livekit-agent"
version = "0.1.0"
description = "LiveKit AI Agent"
requires-python = ">=3.9"
dependencies = [
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector,elevenlabs]~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
    "torch>=2.0.0",
    "PyYAML>=6.0",
    "openai>=1.0.0",
    "elevenlabs>=0.2.0",
    "mutagen>=1.47.0",
    "livekit-plugins-google>=1.1.6",
]

[tool.uv]
dev-dependencies = [] 
